+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add configuration options for subject prefixes (#7929, #4981)
 - Warn the user about a potential data leak on mail bounce or forward (#7993)
 - Make the Empty action available for every non-empty folder, not only Trash (#7948)
+- Remove (incorrect) use of Return-Receipt-To header (#8069)
 - Plugin API: Allow modification of 'error' argument in 'message_send_error' hook (#7914)
 - OAuth: add plugin hooks `oauth_login` and `oauth_refresh_token` for oauth events (#8028, #8040)
 - Debug_logger: Fix the main plugin functionality and documentation (#8041)